requirements.txt
openai
pydantic
rich
python-dotenv